# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/permision/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "permision_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "from": {
                  "example": "any"
                },
                "jabatan_id": {
                  "example": "any"
                },
                "buat_surat": {
                  "example": "any"
                },
                "download_surat": {
                  "example": "any"
                },
                "generate_nomor_surat": {
                  "example": "any"
                },
                "upload_tandatangan": {
                  "example": "any"
                },
                "persetujuan": {
                  "example": "any"
                },
                "tagging": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "jabatan_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "buat_surat": {
                  "example": "any"
                },
                "download_surat": {
                  "example": "any"
                },
                "generate_nomor_surat": {
                  "example": "any"
                },
                "upload_tandatangan": {
                  "example": "any"
                },
                "persetujuan": {
                  "example": "any"
                },
                "tagging": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "permision_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/periode/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "tahun": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "periode_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/periode/status": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "periode_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/daftar-surat/delete": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/daftar-surat/multer/delete/": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perbaikan_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/template-surat/delete": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "template_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/tampilan/delete": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tampilan_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "jabatan_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/status/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "example": "any"
                },
                "from": {
                  "example": "any"
                },
                "surat_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "from": {
                  "example": "any"
                },
                "dibaca": {
                  "example": "any"
                },
                "isSigned": {
                  "example": "any"
                },
                "persetujuan": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "indikator_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/komentar/delete": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "komentar_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/notifikasi/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "from": {
                  "example": "any"
                },
                "surat_id": {
                  "example": "any"
                },
                "jabatan_id_dari": {
                  "example": "any"
                },
                "jabatan_id_ke": {
                  "example": "any"
                },
                "isSign": {
                  "example": "any"
                },
                "persetujuan": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "notifikasi_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "surat_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/sikoja/iku/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "iku_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "iku_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "iku_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/sikoja/indikator/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "indikator_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "strategi_id": {
                  "example": "any"
                },
                "iku_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "indikator_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "strategi_id": {
                  "example": "any"
                },
                "iku_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "indikator_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}